Updates key to avoid usage of index
We cannot assume the suggestion `id` will be unique. This is because at the moment the search results are `Post`s. However in the future we may also need to include `Category` terms and the term IDs could easily clash with the Post IDs as they are in different DB tables.

Using the `type` to differentiate the key.

Addresses https://github.com/WordPress/gutenberg/pull/17846#discussion_r337883174